Module iso7816::aid

Structs

Aid

ISO 7816-4 Application identifier

Enums

Category

FromSliceError

Error returned when the Aid::try_new or Aid::try_new_truncatable fail

Traits

App

According to ISO 7816-4, “Application selection using AID as DF name”: A multi-application card shall support the SELECT command with P1=‘04’, P2=‘00’ and a data field containing 5 to 16 bytes with the AID of an application that may reside on the card. The command shall complete successfully if the AID of an application the card holds matches the data field.